refact: move mapper into persistence module
