+++ peauty-customer-api/src/main/java/com/peauty/customer/implementaion/bidding/EstimateProposalAdapter.java
@@ -4,6 +4,7 @@ import com.peauty.customer.business.bidding.EstimateProposalPort;
 import com.peauty.domain.bidding.EstimateProposal;
 import com.peauty.domain.exception.PeautyException;
 import com.peauty.domain.response.PeautyResponseCode;
+import com.peauty.persistence.bidding.mapper.EstimateProposalMapper;
 import com.peauty.persistence.bidding.estimate.EstimateProposalEntity;
 import com.peauty.persistence.bidding.estimate.EstimateProposalImageEntity;
 import com.peauty.persistence.bidding.estimate.EstimateProposalImageRepository;
