+++ peauty-designer-api/src/main/java/com/peauty/designer/implementation/workspace/WorkspaceAdapter.java
@@ -5,10 +5,11 @@ import com.peauty.domain.designer.Rating;
 import com.peauty.domain.designer.Workspace;
 import com.peauty.domain.exception.PeautyException;
 import com.peauty.domain.response.PeautyResponseCode;
-import com.peauty.persistence.designer.RatingEntity;
-import com.peauty.persistence.designer.RatingRepository;
-import com.peauty.persistence.designer.WorkspaceEntity;
-import com.peauty.persistence.designer.WorkspaceRepository;
+import com.peauty.persistence.designer.mapper.WorkspaceMapper;
+import com.peauty.persistence.designer.rating.RatingEntity;
+import com.peauty.persistence.designer.rating.RatingRepository;
+import com.peauty.persistence.designer.workspace.WorkspaceEntity;
+import com.peauty.persistence.designer.workspace.WorkspaceRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
