+++ peauty-persistence/src/main/java/com/peauty/persistence/designer/badge/BadgeRepository.java
@@ -1,4 +1,4 @@
-package com.peauty.persistence.designer;
+package com.peauty.persistence.designer.badge;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
