+++ peauty-persistence/src/main/java/com/peauty/persistence/designer/workspace/WorkspaceRepository.java
@@ -1,4 +1,4 @@
-package com.peauty.persistence.designer;
+package com.peauty.persistence.designer.workspace;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
